+++ runtime/src/comm/gasnet/comm-gasnet-ex.c
@@ -1834,7 +1834,7 @@ void chpl_rt_comm_execute_on_impl(chpl_rt_prginfo* prg, c_nodeid_t node,
                                   chpl_fn_int_t fid,
                                   chpl_comm_on_bundle_t *arg,
                                   size_t arg_size,
-                                  int ln,
+                                  int64_t ln,
                                   int32_t fn) {
   if (chpl_nodeID == node) {
     assert(0);
@@ -1850,7 +1850,7 @@ void chpl_rt_comm_execute_on_nb_impl(chpl_rt_prginfo* prg, c_nodeid_t node,
                                      chpl_fn_int_t fid,
                                      chpl_comm_on_bundle_t *arg,
                                      size_t arg_size,
-                                     int ln,
+                                     int64_t ln,
                                      int32_t fn) {
   if (chpl_nodeID == node) {
     assert(0); // locale model code should prevent this...
@@ -1867,7 +1867,7 @@ void chpl_rt_comm_execute_on_fast_impl(chpl_rt_prginfo* prg, c_nodeid_t node,
                                        chpl_fn_int_t fid,
                                        chpl_comm_on_bundle_t *arg,
                                        size_t arg_size,
-                                       int ln,
+                                       int64_t ln,
                                        int32_t fn) {
   if (chpl_nodeID == node) {
     assert(0);
